Move debug helpers to debug.py
This makes it possible to run
update_ranges without importing
startup_utils.

Refs #1549
